requirements.txt
python-dateutil==2.8.2
jira==3.4.0
gspread==5.4.0